Add README to sphinx_pyodide and split autodoc_submodules into separate file
